8.9. Оптимизируй это (часть 1)
